use test custom macro for grouping tests together
